refactor: refactor discussions_xblock
