+++ lms/djangoapps/courseware/management/commands/dump_course_structure.py
@@ -25,7 +25,7 @@ from opaque_keys import InvalidKeyError
 from opaque_keys.edx.keys import CourseKey
 from xblock.fields import Scope
 
-from xblock_discussion import DiscussionXBlock
+from xmodule.discussion_block import DiscussionXBlock
 from xmodule.modulestore.django import modulestore
 from xmodule.modulestore.inheritance import compute_inherited_metadata, own_metadata
