+++ openedx/core/djangoapps/discussions/utils.py
@@ -11,7 +11,7 @@ from openedx.core.djangoapps.course_groups.cohorts import get_cohort_names, is_c
 from openedx.core.djangoapps.django_comment_common.models import CourseDiscussionSettings
 from openedx.core.lib.cache_utils import request_cached
 from openedx.core.lib.courses import get_course_by_id
-from openedx.core.lib.xblock_builtin.xblock_discussion.xblock_discussion import DiscussionXBlock
+from xmodule.discussion_block import DiscussionXBlock
 from openedx.core.types import User
 from xmodule.course_module import CourseBlock  # lint-amnesty, pylint: disable=wrong-import-order
 from xmodule.modulestore.django import modulestore  # lint-amnesty, pylint: disable=wrong-import-order
