--- openedx/core/lib/xblock_builtin/xblock_discussion/README.rst
@@ -1,1 +1,0 @@
-See ``lms/djangoapps/discussion/README.rst``
