+++ requirements/edx-sandbox/py38.txt
@@ -14,7 +14,7 @@ click==8.1.3
     #   nltk
 codejail-includes==1.0.0
     # via -r requirements/edx-sandbox/py38.in
-cryptography==37.0.2
+cryptography==37.0.3
     # via -r requirements/edx-sandbox/py38.in
 cycler==0.11.0
     # via matplotlib
