+++ requirements/edx/local.in
@@ -1,5 +1,3 @@
 # Python libraries to install that are local to the edx-platform repo
 -e .
 -e common/lib/capa
-
--e openedx/core/lib/xblock_builtin/xblock_discussion
