+++ xmodule/discussion_block.py
@@ -14,12 +14,10 @@ from xblock.core import XBlock
 from xblock.fields import UNIQUE_ID, Scope, String
 from xblockutils.resources import ResourceLoader
 from xblockutils.studio_editable import StudioEditableXBlockMixin
-from xmodule.xml_module import XmlParserMixin
 
-from lms.djangoapps.discussion.toggles import ENABLE_DISCUSSIONS_MFE
-from openedx.core.djangoapps.discussions.url_helpers import get_discussions_mfe_topic_url
 from openedx.core.djangolib.markup import HTML, Text
-from openedx.core.lib.xblock_builtin import get_css_dependencies, get_js_dependencies
+from openedx.core.lib.xblock_utils import get_css_dependencies, get_js_dependencies
+from xmodule.xml_module import XmlParserMixin
 
 log = logging.getLogger(__name__)
 loader = ResourceLoader(__name__)  # pylint: disable=invalid-name
